clustered map
I added a new map component that clusters sites if they're close enough together. Right now I have both maps (clustered and unclustered) in database.vue. The clustered map does seem to be a bit faster but I still prefer the unclustered map because it's not that slow and it gives you a better sense of where things really are