Router updated to support the orcid login.
Making a user redirection page to help ensure that things are working. When a user logs in and is redirected back from ORCID they are pushed off to the user page.

This is done through the `removeORCIDHash()` function in the router file. ORCID redirects to `login`, and then passes the users over to users, after storing the user information in localstorage.