Working on the auth process, trying to log for bugs.
I've added some coments to the head of the auth file to explain the process.

Right now there seems to be an issue with re-loading a user once they've logged in. The access token only is valid for one use. So we want to store the information about the user, but then be able to pull that in at any time.